refactor: keep collectionContent controller-free for application-layer use
