+++ src/components/organisms/Collections/CollectionCard/CollectionCard.tsx
@@ -17,7 +17,8 @@ import { useIsMobile } from '@/hooks/useIsMobile/useIsMobile';
 import { usePostDetails } from '@/hooks/usePostDetails/usePostDetails';
 import { useRequireAuth } from '@/hooks/useRequireAuth/useRequireAuth';
 import { useUserProfile } from '@/hooks/useUserProfile/useUserProfile';
-import { parseCollectionContent, resolveCollectionCoverImage } from '@/libs/post/collectionContent';
+import { parseCollectionContent } from '@/libs/post/collectionContent';
+import { resolveCollectionCoverImage } from '@/libs/post/collectionCoverImage';
 import { cn, isPostDeleted } from '@/libs/utils/utils';
 import type { Pubky } from '@/models/models.types';
 import { buildCompositeId } from '@/models/models.utils';
