+++ src/components/organisms/Collections/CollectionHero/CollectionHero.tsx
@@ -14,7 +14,8 @@ import { useDeletePost } from '@/hooks/useDeletePost/useDeletePost';
 import { usePostReplyRepostDialogs } from '@/hooks/usePostReplyRepostDialogs/usePostReplyRepostDialogs';
 import { useRequireAuth } from '@/hooks/useRequireAuth/useRequireAuth';
 import { useUserProfile } from '@/hooks/useUserProfile/useUserProfile';
-import { parseCollectionContent, resolveCollectionCoverImage } from '@/libs/post/collectionContent';
+import { parseCollectionContent } from '@/libs/post/collectionContent';
+import { resolveCollectionCoverImage } from '@/libs/post/collectionCoverImage';
 import { cn } from '@/libs/utils/utils';
 import { buildCompositeId } from '@/models/models.utils';
 import { CollectionCountBadge } from '@/molecules/CollectionCountBadge/CollectionCountBadge';
