+++ src/hooks/useEditCollection/useEditCollection.ts
@@ -16,7 +16,8 @@ import { usePostDetails } from '@/hooks/usePostDetails/usePostDetails';
 import { isAppError } from '@/libs/error/error.utils';
 import { getImageUploadSizeLimitToastMessage } from '@/libs/image/imageUploadSizeLimit';
 import { Logger } from '@/libs/logger/logger';
-import { parseCollectionContent, resolveCollectionCoverImage } from '@/libs/post/collectionContent';
+import { parseCollectionContent } from '@/libs/post/collectionContent';
+import { resolveCollectionCoverImage } from '@/libs/post/collectionCoverImage';
 import { useToast } from '@/molecules/Toaster/use-toast';
 import { FileVariant } from '@/services/nexus/file/file.types';
 import { useLocalFilesStore } from '@/stores/localFiles/localFiles.store';
